+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow to set PYOUPYOU_ENV from a file
 - Fix create process form (remove subscribe to notification field)
 - Strip search
+- Fix gantt chart when no data
 
 # v1.19.0 (2023-03-16)
 - Allow user to subscribe to a recruitment process